CastIt.GoogleCast.Cli was renamed to CastIt.Cli
