+++ .gitignore
@@ -4,8 +4,8 @@
 ## Get latest from https://github.com/github/gitignore/blob/master/VisualStudio.gitignore
 
 CastIt.Setup/Installers/**
-CastIt.GoogleCast.Cli.Setup/Installers/**
-CastIt.GoogleCast.Cli.Setup/Source/**
+CastIt.Cli.Setup/Installers/**
+CastIt.Cli.Setup/Source/**
 Secrets.cs
 *.g.dart
 *.freezed.dart
