+++ CastIt.Cli.Setup/Assets/License.rtf
