+++ CastIt.Cli.Setup/CliComponents.wxs
