+++ CastIt.Cli.Setup/Directories.wxs
