+++ CastIt.Cli.Setup/Product.wxs
