+++ CastIt.Cli/CastIt.Cli.csproj
@@ -5,6 +5,7 @@
     <TargetFramework>net5.0</TargetFramework>
     <!--The name here matters, we set it like this so we can type castit in the cmd-->
     <AssemblyName>CastIt</AssemblyName>
+    <RootNamespace>CastIt.Cli</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
