+++ CastIt.Cli/Commands/BaseCommand.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
 
-namespace CastIt.GoogleCast.Cli.Commands
+namespace CastIt.Cli.Commands
 {
     [HelpOption("--help", ShortName = "h")]
     public abstract class BaseCommand
