+++ CastIt.Cli/Commands/ConnectCommand.cs
@@ -1,13 +1,13 @@
 ﻿using CastIt.Application.Server;
-using CastIt.GoogleCast.Cli.Common.Utils;
-using CastIt.GoogleCast.Cli.Interfaces.Api;
+using CastIt.Cli.Common.Utils;
+using CastIt.Cli.Interfaces.Api;
 using McMaster.Extensions.CommandLineUtils;
 using Refit;
 using System;
 using System.Net;
 using System.Threading.Tasks;
 
-namespace CastIt.GoogleCast.Cli.Commands
+namespace CastIt.Cli.Commands
 {
     [Command(Name = "connect", Description = "Connects to a particular device")]
     public class ConnectCommand : BaseCommand
