+++ CastIt.Cli/Commands/StopCommand.cs
@@ -1,25 +1,19 @@
 ﻿using CastIt.Application.Server;
-using CastIt.GoogleCast.Cli.Common.Utils;
-using CastIt.GoogleCast.Cli.Interfaces.Api;
+using CastIt.Cli.Common.Utils;
+using CastIt.Cli.Interfaces.Api;
 using McMaster.Extensions.CommandLineUtils;
 using Refit;
 using System;
 using System.Threading.Tasks;
 
-namespace CastIt.GoogleCast.Cli.Commands
+namespace CastIt.Cli.Commands
 {
-    [Command(Name = "volume", Description = "Sets the volume of the device")]
-    public class VolumeCommand : BaseCommand
+    [Command(Name = "stop", Description = "Stops the playback of the current played file")]
+    public class StopCommand : BaseCommand
     {
         private readonly IConsole _console;
 
-        [Option(CommandOptionType.NoValue, Description = "Pass this option to mute the device.", LongName = "muted")]
-        public bool IsMuted { get; set; }
-
-        [Option(CommandOptionType.SingleOrNoValue, Description = "The volume level of the device, it has to be a value between 0 - 100. Defaults to 100", LongName = "level")]
-        public float Volume { get; set; } = 100;
-
-        public VolumeCommand(IConsole console)
+        public StopCommand(IConsole console)
         {
             _console = console;
         }
@@ -34,17 +28,17 @@ namespace CastIt.GoogleCast.Cli.Commands
                     return -1;
                 }
 
-                _console.WriteLine($"Setting volume to = {Volume} and muted = {IsMuted}...");
+                _console.WriteLine("Stopping playback of current played file...");
                 var url = ServerUtils.StartServerIfNotStarted(_console);
                 var castItApi = RestService.For<ICastItApi>(url);
-                var response = await castItApi.SetVolume(Volume, IsMuted);
+                var response = await castItApi.Stop();
                 if (!response.Succeed)
                 {
                     _console.WriteLine(response.Message);
                     return -1;
                 }
 
-                _console.WriteLine("Volume was successfully updated");
+                _console.WriteLine("Playback was successfully stopped");
             }
             catch (Exception e)
             {
