+++ CastIt.Cli/Properties/launchSettings.json
