+++ CastIt.Cli/appsettings.json
