+++ CastIt.sln
@@ -11,8 +11,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CastIt.GoogleCast", "CastIt
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CastIt.Setup.CustomActions", "CastIt.Setup.CustomActions\CastIt.Setup.CustomActions.csproj", "{9771D1A9-11EC-4104-B537-25BB9EC3BAF5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CastIt.GoogleCast.Cli", "CastIt.GoogleCast.Cli\CastIt.GoogleCast.Cli.csproj", "{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CastIt.Server", "CastIt.Server\CastIt.Server.csproj", "{4DD24030-B53E-4D2F-8CCC-362203BE1AAA}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CastIt.Domain", "CastIt.Domain\CastIt.Domain.csproj", "{E060BBBC-BBFF-4F24-82BF-4829B686CC00}"
@@ -23,9 +21,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CastIt.Shared", "CastIt.Sha
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CastIt.Infrastructure", "CastIt.Infrastructure\CastIt.Infrastructure.csproj", "{DF7A861A-F8A6-4D49-8D10-8AB2E4999D2C}"
 EndProject
-Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "CastIt.GoogleCast.Cli.Setup", "CastIt.GoogleCast.Cli.Setup\CastIt.GoogleCast.Cli.Setup.wixproj", "{C8841EB1-E5AE-44D2-B05E-AECA86542E94}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Setups", "Setups", "{BC7DAE32-56A5-47EC-811A-D0861DD01FFD}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CastIt.Cli", "CastIt.Cli\CastIt.Cli.csproj", "{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}"
+EndProject
+Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "CastIt.Cli.Setup", "CastIt.Cli.Setup\CastIt.Cli.Setup.wixproj", "{C8841EB1-E5AE-44D2-B05E-AECA86542E94}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -79,18 +79,6 @@ Global
 		{9771D1A9-11EC-4104-B537-25BB9EC3BAF5}.Release|x64.Build.0 = Release|x64
 		{9771D1A9-11EC-4104-B537-25BB9EC3BAF5}.Release|x86.ActiveCfg = Release|x86
 		{9771D1A9-11EC-4104-B537-25BB9EC3BAF5}.Release|x86.Build.0 = Release|x86
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Debug|x64.Build.0 = Debug|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Debug|x86.Build.0 = Debug|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Release|x64.ActiveCfg = Release|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Release|x64.Build.0 = Release|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Release|x86.ActiveCfg = Release|Any CPU
-		{0340CCF7-D1C1-4DDC-B836-3E652C6E6056}.Release|x86.Build.0 = Release|Any CPU
 		{4DD24030-B53E-4D2F-8CCC-362203BE1AAA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4DD24030-B53E-4D2F-8CCC-362203BE1AAA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4DD24030-B53E-4D2F-8CCC-362203BE1AAA}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -151,6 +139,18 @@ Global
 		{DF7A861A-F8A6-4D49-8D10-8AB2E4999D2C}.Release|x64.Build.0 = Release|Any CPU
 		{DF7A861A-F8A6-4D49-8D10-8AB2E4999D2C}.Release|x86.ActiveCfg = Release|Any CPU
 		{DF7A861A-F8A6-4D49-8D10-8AB2E4999D2C}.Release|x86.Build.0 = Release|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Debug|x64.Build.0 = Debug|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Debug|x86.Build.0 = Debug|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Release|Any CPU.Build.0 = Release|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Release|x64.ActiveCfg = Release|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Release|x64.Build.0 = Release|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Release|x86.ActiveCfg = Release|Any CPU
+		{99B5C9DB-E0FA-4A36-9CFC-2F2755861F14}.Release|x86.Build.0 = Release|Any CPU
 		{C8841EB1-E5AE-44D2-B05E-AECA86542E94}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{C8841EB1-E5AE-44D2-B05E-AECA86542E94}.Debug|x64.ActiveCfg = Debug|x64
 		{C8841EB1-E5AE-44D2-B05E-AECA86542E94}.Debug|x64.Build.0 = Debug|x64
